refact: split api/ark into multiple files
